WIP: efi: setup GRUB_ENV EFI variable
Grub evaluates the contents of the GRUB_ENV variable as a envblk.  So
this can be used as a sneaky way to pass $fdtfile or other variables to
grub.

(Probably a cleaner, less distro specific, solution would be to give
grub a command to read EFI variables into environment variables.  But
grub2 doesn't have that yet.)

Note that I tried to make this part of the boot_efi_binary bootenv,
but the contents of GRUB_ENV are whitespace sensitive, and the bootenv
parsing doesn't appear to let me escape quotes.